# public/openapi.yaml
openapi: 3.0.3
info:
  title: EXA-TA-WEB2 Academic Blockchain Network API
  version: 1.0.0
  description: >
    Distributed blockchain system where each node manages academic degrees,
    propagates transactions, mines blocks, and reaches consensus.
servers:
  - url: http://10.173.212.x:8001/api
...
components:
  schemas:
...
    Block:
      type: object
      required:
        - persona_id
        - institucion_id
        - programa_id
        - fecha_fin
        - titulo_obtenido
        - hash_actual
        - hash_anterior
        - nonce
      properties:
        persona_id:
          type: string
        institucion_id:
          type: string
        programa_id:
          type: string
        fecha_inicio:
          type: string
          format: date
        fecha_fin:
          type: string
          format: date
        titulo_obtenido:
          type: string
        numero_cedula:
          type: string
        titulo_tesis:
          type: string
        menciones:
          type: string
        firmado_por:
          type: string
        hash_actual:
          type: string
        hash_anterior:
          type: string
        nonce:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string

    # DATOS BASE 
...
paths:
...
  /blocks:
    post:
      tags: [Blocks]
      summary: Receive a mined block from another node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [block]
              properties:
                block:
                  $ref: '#/components/schemas/Block'
      responses:
        '200':
          description: Block accepted or already synced
        '201':
          description: Block accepted and saved
        '409':
          description: Invalid block, desync detected. Triggers auto-consensus.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...